Rename remoteErrorReducer to errorsReducer
[INT-35]
