+++ app/javascript/reducers/errorsReducer.js
