+++ spec/javascripts/store/storeSpec.js
@@ -27,6 +27,7 @@ describe('Store', () => {
       counties: [],
       countyAgencies: [],
       crossReportForm: {},
+      errors: {},
       incidentInformationForm: {},
       investigation: {},
       involvements: [],
@@ -35,7 +36,6 @@ describe('Store', () => {
       participants: [],
       peopleForm: {},
       relationships: [],
-      remoteError: {},
       routing: {locationBeforeTransitions: null},
       screening: {},
       screeningInformationForm: {},
